+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/domain/case-flag.model.ts
@@ -8,16 +8,22 @@ export interface FlagPath {
 export interface FlagDetail {
   id?: string;
   name: string;
+  name_cy?: string;
   subTypeValue?: string;
+  subTypeValue_cy?: string;
   subTypeKey?: string;
   otherDescription?: string;
+  otherDescription_cy?: string;
   flagComment?: string;
+  flagComment_cy?: string;
+  flagUpdateComment?: string;
   dateTimeModified?: Date | string;
   dateTimeCreated: Date | string;
   path: FlagPath[];
   hearingRelevant: boolean | string;
   flagCode: string;
   status: string;
+  availableExternally?: boolean | string;
 }
 
 export interface Flags {