+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/enums/case-flag-wizard-step-title.enum.ts
@@ -6,6 +6,7 @@ export enum CaseFlagWizardStepTitle {
   SEARCH_SIGN_LANGUAGE_INTERPRETER = 'Search for a sign language interpreter',
   ADD_FLAG_COMMENTS = 'Add comments for this flag',
   ADD_FLAG_COMMENTS_EXTERNAL_MODE = 'Tell us more about the request',
+  CONFIRM_FLAG_STATUS = 'Confirm the status of the flag',
   FLAG_STATUS = 'Flag status',
   MANAGE_CASE_FLAGS = 'Manage case flags',
   UPDATE_FLAG_TITLE = 'Update flag'